moved CsrfDisabledCheck into /hotspots
