+++ python-checks/src/main/java/org/sonar/python/checks/CheckList.java
@@ -25,6 +25,7 @@ import java.util.HashSet;
 import org.sonar.python.checks.hotspots.ClearTextProtocolsCheck;
 import org.sonar.python.checks.hotspots.CommandLineArgsCheck;
 import org.sonar.python.checks.hotspots.CorsCheck;
+import org.sonar.python.checks.hotspots.CsrfDisabledCheck;
 import org.sonar.python.checks.hotspots.DataEncryptionCheck;
 import org.sonar.python.checks.hotspots.DebugModeCheck;
 import org.sonar.python.checks.hotspots.DisabledHtmlAutoEscapeCheck;
