+++ python-checks/src/test/resources/checks/hotspots/csrfDisabledCheck/django/views.py
