+++ python-checks/src/test/resources/checks/hotspots/csrfDisabledCheck/flask/exemptAsFunction.py
