+++ python-checks/src/test/resources/checks/hotspots/csrfDisabledCheck/flask/flaskExempt.py
