+++ python-checks/src/test/resources/checks/hotspots/csrfDisabledCheck/flask/flaskform1.py
