+++ python-checks/src/test/resources/checks/hotspots/csrfDisabledCheck/flask/wtfCsrfEnabledFalse.py
